backend/filtering-service-b/pyproject.toml
[project]
name = "filtering-service-b"
version = "0.1.0"
description = "Filtering Service B Phase 1 (Kafka IO skeleton)"
authors = [{ name = "wasiflatifhussain" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.12.2,<3.0.0",
    "python-dotenv>=1.2.1,<2.0.0",
    "fastapi>=0.127.0,<0.128.0",
    "uvicorn>=0.40.0,<0.41.0",
    "redis>=5.2.1,<6.0.0",
    "sentence-transformers (>=5.3.0,<6.0.0)",
]

[project.optional-dependencies]
dev = ["pytest>=8.4.0,<9.0.0"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "filtering_service_b", from = "src" }]
